[explorer/frontend] feat: relocate NEM API into variant behind thin re-exports
