+++ explorer/frontend/variants/nem/api/health.js
@@ -1,0 +1,5 @@
+import { createApiUrl, makeRequest } from '@/app/utils/server';
+
+export const fetchBackendHealthStatus = async () => {
+	return makeRequest(createApiUrl('health'));
+};
