+++ explorer/frontend/variants/nem/api/index.js
@@ -1,0 +1,9 @@
+export * as accounts from './accounts';
+export * as blocks from './blocks';
+export * as health from './health';
+export * as mosaics from './mosaics';
+export * as namespaces from './namespaces';
+export * as nodes from './nodes';
+export * as search from './search';
+export * as stats from './stats';
+export * as transactions from './transactions';
