+++ explorer/frontend/variants/nem/index.js
@@ -1,0 +1,6 @@
+export * as api from './api';
+export { DocumentHead } from './DocumentHead';
+export { default as styleVariables } from './styles/variables.json';
+export { default as pageConfig } from './config/pages';
+export { default as components } from './components';
+export { default as config } from './config';
